Move theme utils under /utils
